refactor: rename files for clarity per PR feedback
- templates.py → jinja2_templates.py (clarifies it's for Jinja2)
- frontend.py → eslint_quick.py (clarifies it's ESLint quick check)
- Updated all imports and test files
