+++ slopbucket/checks/javascript/eslint_quick.py
